refactor: 입장코드 유효성 확인 유틸 함수 분리
